Move winners logic into winners service
